Create Spotify Playlist Doubly Linked List Implementation.cpp
This program demonstrates the implementation of a doubly linked list to model a Spotify playlist, a concept learned in Data Structures and Algorithms as part of an undergraduate degree in Information Systems and Applied Computing. Each song is represented by a SongNode that stores the song title along with two pointers: one pointing to the next song and another pointing to the previous song. This two-way linkage allows the playlist to be traversed both forward (next song) and backward (previous song), similar to how real music apps support next and previous navigation. The InsertAfter function maintains proper links between nodes, while PushHead uses a pointer-to-pointer technique to safely update the head of the list when new songs are added to the front.